c# update unit_text
